refactor visualization as standalone component
