+++ python/Makefile.am
@@ -184,10 +184,10 @@ endif # MAINTAINER_MODE
 # specification of python source files to be byte-compiled at installation
 ######################################################################
 HL_IFACE =                  \
-    $(srcdir)/animation.py       \
     $(srcdir)/geom.py       \
     $(srcdir)/simulation.py \
     $(srcdir)/source.py     \
+    $(srcdir)/visualization.py  \
     $(srcdir)/materials.py
 
 pkgpython_PYTHON = __init__.py $(HL_IFACE)
